backend/requirements.txt
boto3==1.16.25
botocore==1.19.25
Pillow==8.0.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
requests==2.21.0
bcrypt==3.1.4
click==6.7
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.1.1
pycparser==2.19
six==1.14.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1